Make wallet relation polymorphic

+++ database/migrations/2018_02_20_113500_create_wallet_transactions_table.php
@@ -17,7 +17,7 @@ class CreateWalletTransactionsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('wallet_id');
 
-            $table->integer('amount'); // amount is an integer, it could be "dollars" or "cents"
+            $table->decimal('amount', 12, 4); // amount is an integer, it could be "dollars" or "cents"
             $table->string('hash', 60); // hash is a uniqid for each transaction
             $table->string('type', 30); // type can be anything in your app, by default we use "deposit" and "withdraw"
             $table->boolean('accepted'); // All transactions will be added in the book, some can be refused
